toke.c: Extract part of \N{} processing to new function
This is in preparation for making the deprecations that this code covers
and making them fatal.

This code combines the first and final portions of the code that handles
\N{names}, leaving the middle intact.  There are no intentional logic
changes.  The code is moved and outdented as appropriate for not being
within nested 'if's, and the comments are reflowed to fill 79 columns.
One declaration had a const added.

This is needed to be done, as the current warnings become fatal, and
should be done before processing is wasted on computing the values that
are about to be thrown away.